test(http): split routes.test.ts and rename chat.test.ts
- chat.test.ts → sessions.test.ts (matches sessions.ts)
- routes.test.ts split into cron.test.ts + status.test.ts;
  the two orphan session tests fold into sessions.test.ts
- Extract shared startTestServer + request helpers to test-helpers.ts

Every test file now maps 1:1 to its production file.
